fixed cwl_parser module and its test
